scripts/config/requirements.txt
python-can
simple-term-menu